final javascript/css commit
Added a width of 50%,30% to the article and aside. Displays the information in the variable ratingReport inside of the aside in the innerHTML. Added the summaries of the author under the image and will be on the page.  putting the star image to the page html. Each time through the for loop, the author will be rated based on a star scale. The for loop is set to 0 to 2 and the loop will only show the first three reviews of the authors. The initial text string is set to the html and here average is the var rating avg is concatenated to the inner html. The rating Avg is equal to the rating sum divided by the ratingsCount values. Information about the game that is displayed on the page. The title, manufacturer, id, price, platform, esrb, condition have the corresponding variables that are concatenated with the information. Each time through the loop, the value of the current ratings value is added to the value of the ratings sum var. Ready for review.